# perf: Offer 'simd' feature for faster folding 