--- a/Cargo.toml
+++ b/Cargo.toml
@@ -29,8 +29,8 @@ unicode-width = "0.1.11"
 
 [dev-dependencies]
 anstream = "0.6.13"
-criterion = "0.5.1"
 difference = "2.0.0"
+divan = "0.1.14"
 glob = "0.3.1"
 serde = { version = "1.0.199", features = ["derive"] }
 snapbox = { version = "0.6.0", features = ["diff", "term-svg", "cmd", "examples"] }